requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.55.0
python-multipart>=0.0.12
pypdf>=5.0.0
python-dotenv>=1.0.0
